changed health to discrete, changed bullets damage, added half health, changed player movement

+++ Assets/Scripts/Player/Bullet.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-    private float damage;
-    public void SetDamage(float dmg)
+    private int damage;
+    public void SetDamage(int dmg)
     {
         damage = dmg;
     }
